remove duplicate grid step
